third_party/blink/web_tests/platform/mac/external/wpt/encoding/unsupported-labels.window-expected.txt

Content-Type: text/plain
This is a testharness.js-based test.
Found 158 tests; 157 PASS, 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS 437 is not supported by the Encoding Standard
PASS adobe-standard-encoding is not supported by the Encoding Standard
PASS armscii-8 is not supported by the Encoding Standard
PASS bocu-1 is not supported by the Encoding Standard
PASS cesu-8 is not supported by the Encoding Standard
PASS cp1025 is not supported by the Encoding Standard
PASS cp437 is not supported by the Encoding Standard
PASS cp737 is not supported by the Encoding Standard
PASS cp851 is not supported by the Encoding Standard
PASS cp858 is not supported by the Encoding Standard
PASS cp862 is not supported by the Encoding Standard
PASS cp864 is not supported by the Encoding Standard
PASS cp869 is not supported by the Encoding Standard
PASS cp875 is not supported by the Encoding Standard
PASS cp950 is not supported by the Encoding Standard
PASS csiso103t618bit is not supported by the Encoding Standard
PASS csiso111ecmacyrillic is not supported by the Encoding Standard
PASS cspc8codepage437 is not supported by the Encoding Standard
PASS csviscii is not supported by the Encoding Standard
PASS dos-720 is not supported by the Encoding Standard
PASS dos-862 is not supported by the Encoding Standard
PASS ecma-cyrillic is not supported by the Encoding Standard
PASS euc-tw is not supported by the Encoding Standard
PASS german is not supported by the Encoding Standard
PASS geostd8 is not supported by the Encoding Standard
PASS hp-roman8 is not supported by the Encoding Standard
PASS ibm-thai is not supported by the Encoding Standard
PASS ibm00858 is not supported by the Encoding Standard
PASS ibm00924 is not supported by the Encoding Standard
PASS ibm01047 is not supported by the Encoding Standard
PASS ibm01140 is not supported by the Encoding Standard
PASS ibm01141 is not supported by the Encoding Standard
PASS ibm01142 is not supported by the Encoding Standard
PASS ibm01143 is not supported by the Encoding Standard
PASS ibm01144 is not supported by the Encoding Standard
PASS ibm01145 is not supported by the Encoding Standard
PASS ibm01146 is not supported by the Encoding Standard
PASS ibm01147 is not supported by the Encoding Standard
PASS ibm01148 is not supported by the Encoding Standard
PASS ibm01149 is not supported by the Encoding Standard
PASS ibm037 is not supported by the Encoding Standard
PASS ibm1026 is not supported by the Encoding Standard
PASS ibm1047 is not supported by the Encoding Standard
PASS ibm273 is not supported by the Encoding Standard
PASS ibm277 is not supported by the Encoding Standard
PASS ibm278 is not supported by the Encoding Standard
PASS ibm280 is not supported by the Encoding Standard
PASS ibm284 is not supported by the Encoding Standard
PASS ibm285 is not supported by the Encoding Standard
PASS ibm290 is not supported by the Encoding Standard
PASS ibm297 is not supported by the Encoding Standard
PASS ibm367 is not supported by the Encoding Standard
PASS ibm420 is not supported by the Encoding Standard
PASS ibm423 is not supported by the Encoding Standard
PASS ibm424 is not supported by the Encoding Standard
PASS ibm437 is not supported by the Encoding Standard
PASS ibm500 is not supported by the Encoding Standard
PASS ibm737 is not supported by the Encoding Standard
PASS ibm775 is not supported by the Encoding Standard
PASS ibm850 is not supported by the Encoding Standard
PASS ibm852 is not supported by the Encoding Standard
PASS ibm855 is not supported by the Encoding Standard
PASS ibm857 is not supported by the Encoding Standard
PASS ibm860 is not supported by the Encoding Standard
PASS ibm861 is not supported by the Encoding Standard
PASS ibm862 is not supported by the Encoding Standard
PASS ibm863 is not supported by the Encoding Standard
PASS ibm864 is not supported by the Encoding Standard
PASS ibm864i is not supported by the Encoding Standard
PASS ibm865 is not supported by the Encoding Standard
PASS ibm868 is not supported by the Encoding Standard
PASS ibm869 is not supported by the Encoding Standard
PASS ibm870 is not supported by the Encoding Standard
PASS ibm871 is not supported by the Encoding Standard
PASS ibm880 is not supported by the Encoding Standard
PASS ibm905 is not supported by the Encoding Standard
PASS ibm918 is not supported by the Encoding Standard
PASS iso-2022-jp-1 is not supported by the Encoding Standard
PASS iso-2022-jp-2 is not supported by the Encoding Standard
PASS iso-2022-jp-3 is not supported by the Encoding Standard
PASS iso-8859-8 visual is not supported by the Encoding Standard
PASS jis_c6226-1978 is not supported by the Encoding Standard
PASS jis_x0208-1983 is not supported by the Encoding Standard
PASS jis_x0208-1990 is not supported by the Encoding Standard
PASS jis_x0212-1990 is not supported by the Encoding Standard
PASS johab is not supported by the Encoding Standard
PASS latin9 is not supported by the Encoding Standard
PASS norwegian is not supported by the Encoding Standard
PASS sami-ws2 is not supported by the Encoding Standard
PASS scsu is not supported by the Encoding Standard
PASS shift_jis_x0213-2000 is not supported by the Encoding Standard
PASS swedish is not supported by the Encoding Standard
PASS tcvn is not supported by the Encoding Standard
PASS tis-620-2533 is not supported by the Encoding Standard
PASS utf-7 is not supported by the Encoding Standard
PASS utf-32 is not supported by the Encoding Standard
PASS viscii is not supported by the Encoding Standard
FAIL windows-936-2000 is not supported by the Encoding Standard assert_equals: expected "windows-1252" but got "GBK"
PASS windows-sami-2 is not supported by the Encoding Standard
PASS ws2 is not supported by the Encoding Standard
PASS x-chinese-cns is not supported by the Encoding Standard
PASS x-chinese-eten is not supported by the Encoding Standard
PASS x-cp20001 is not supported by the Encoding Standard
PASS x-cp20003 is not supported by the Encoding Standard
PASS x-cp20004 is not supported by the Encoding Standard
PASS x-cp20005 is not supported by the Encoding Standard
PASS x-cp20261 is not supported by the Encoding Standard
PASS x-cp20269 is not supported by the Encoding Standard
PASS x-cp20936 is not supported by the Encoding Standard
PASS x-cp20949 is not supported by the Encoding Standard
PASS x-cp21027 is not supported by the Encoding Standard
PASS x-cp50227 is not supported by the Encoding Standard
PASS x-cp50229 is not supported by the Encoding Standard
PASS x-ebcdic-koreanextended is not supported by the Encoding Standard
PASS x-europa is not supported by the Encoding Standard
PASS x-ia5 is not supported by the Encoding Standard
PASS x-ia5-german is not supported by the Encoding Standard
PASS x-ia5-norwegian is not supported by the Encoding Standard
PASS x-ia5-swedish is not supported by the Encoding Standard
PASS x-iscii-as is not supported by the Encoding Standard
PASS x-iscii-be is not supported by the Encoding Standard
PASS x-iscii-de is not supported by the Encoding Standard
PASS x-iscii-gu is not supported by the Encoding Standard
PASS x-iscii-ka is not supported by the Encoding Standard
PASS x-iscii-ma is not supported by the Encoding Standard
PASS x-iscii-or is not supported by the Encoding Standard
PASS x-iscii-pa is not supported by the Encoding Standard
PASS x-iscii-t is not supported by the Encoding Standard
PASS x-iscii-ta is not supported by the Encoding Standard
PASS x-iscii-te is not supported by the Encoding Standard
PASS x-mac-arabic is not supported by the Encoding Standard
PASS x-mac-ce is not supported by the Encoding Standard
PASS x-mac-centraleurroman is not supported by the Encoding Standard
PASS x-mac-chinesesimp is not supported by the Encoding Standard
PASS x-mac-chinesetrad is not supported by the Encoding Standard
PASS x-mac-croatian is not supported by the Encoding Standard
PASS x-mac-devanagari is not supported by the Encoding Standard
PASS x-mac-dingbats is not supported by the Encoding Standard
PASS x-mac-farsi is not supported by the Encoding Standard
PASS x-mac-greek is not supported by the Encoding Standard
PASS x-mac-gujarati is not supported by the Encoding Standard
PASS x-mac-gurmukhi is not supported by the Encoding Standard
PASS x-mac-hebrew is not supported by the Encoding Standard
PASS x-mac-icelandic is not supported by the Encoding Standard
PASS x-mac-japanese is not supported by the Encoding Standard
PASS x-mac-korean is not supported by the Encoding Standard
PASS x-mac-roman-latin1 is not supported by the Encoding Standard
PASS x-mac-romanian is not supported by the Encoding Standard
PASS x-mac-symbol is not supported by the Encoding Standard
PASS x-mac-thai is not supported by the Encoding Standard
PASS x-mac-tibetan is not supported by the Encoding Standard
PASS x-mac-turkish is not supported by the Encoding Standard
PASS x-mac-vt100 is not supported by the Encoding Standard
PASS x-nextstep is not supported by the Encoding Standard
PASS x-vps is not supported by the Encoding Standard
PASS _autodetect is not supported by the Encoding Standard
PASS _autodetect_all is not supported by the Encoding Standard
PASS _autodetect_kr is not supported by the Encoding Standard
Harness: the test ran to completion.

#EOF
#EOF
